+++ Commands/Admin.js
@@ -2,7 +2,8 @@ const Discord = require('discord.js')
 const { spawn } = require('child_process');
 const fh = require('./FileHandler');
 const { version } = require('../package.json');
-const Logger = require('./Logger.js');
+const Logger = require('./Logger');
+const Channel = require('./Channel');
 
 /**
  * @usage uwuadmin <command>
@@ -31,58 +32,83 @@ module.exports = {
     },
 
     /**
-     * @summary Checks if User that called an AdminCommand is an Admin, is useless if called in Discord
+     * @summary Checks if User that called an AdminCommand is an Admin/Developer, is useless if called in Discord
      * @returns boolean
      */
     isAdmin: function (message) {
-        return Admins.includes(message.author.id);
+        if (this.isDev(message)) {
+            return true;
+        }
+        return fh.get('../Files/local/' + message.guild.id + '/Admins.json').includes(message.author.id); //To be replaced with perServer-Check
+    },
+
+    /**
+     * @summary Checks if User that called an AdminCommand is a Developer, is useless if called in Discord
+     * @returns boolean
+     */
+    isDev: function (message) {
+        return Devs.includes(message.author.id);
     },
 
     /**
      * @summary Updates the Bot to the newest version on github, will restart the Bot so LogFile is lost
+     * @DevOnly
      */
-    update: function (message, bot) {
-        message.channel.send("Updating now").then(_ => {
+    update: function (message) {
+        if (this.isDev(message)) {
+            message.channel.send("Updating now").then(_ => {
 
-            spawn('start', ['cmd.exe', '/c', '.\\Files\\Updater.bat'], { shell: true })
-                .on('exit', _ => {
-                    process.exit(0);
-                })
-        });
+                spawn('start', ['cmd.exe', '/c', '.\\Files\\Updater.bat'], { shell: true })
+                    .on('exit', _ => {
+                        process.exit(0);
+                    })
+            });
+        } else {
+            message.channel.send('You are not a Developer');
+        }
     },
 
     /**
      * @summary Stops the Bot if called twice within 10 Seconds
+     * @DevOnly
      */
     stop: function (message, bot) {
+        if (this.isDev(message)) {
+            if (stopvar) {
 
-        if (stopvar) {
+                message.channel.send("Stopping now").then(_ => {
+                    bot.user.setPresence({ game: { name: 'on ' + version }, status: 'offline' }).then(_ => {
+                        process.exit(0);
+                    })
+                });
 
-            message.channel.send("Stopping now").then(_ => {
-                bot.user.setPresence({ game: { name: 'on ' + version }, status: 'offline' }).then(_ => {
-                    process.exit(0);
-                })
-            });
+            } else {
 
+                message.channel.send("If you really want to stop the Bot call this function again within 10 sec");
+                stopvar = true;
+                setTimeout(function () { stopvar = false; }, 10000);
+            }
         } else {
-
-            message.channel.send("If you really want to stop the Bot call this function again within 10 sec");
-            stopvar = true;
-            setTimeout(function () { stopvar = false; }, 10000);
+            message.channel.send('You are not a Developer');
         }
     },
 
     /**
      * @summary Restarts the Bot, will delete the LogFile until now so be careful
+     * @DevOnly
      */
     restart: function (message) {
-        message.channel.send('Restarting now').then(_ => {
+        if (this.isDev(message)) {
+            message.channel.send('Restarting now').then(_ => {
 
-            spawn('start', ['cmd.exe', '/c', 'run.bat'], { shell: true })
-                .on('exit', m => {
-                    process.exit(0);
-                })
-        })
+                spawn('start', ['cmd.exe', '/c', 'run.bat'], { shell: true })
+                    .on('exit', m => {
+                        process.exit(0);
+                    })
+            })
+        } else {
+            message.channel.send('You are not a Developer');
+        }
     },
 
     /**
@@ -96,7 +122,7 @@ module.exports = {
     * @summary sets the LeagueChannel to the current Channel
     */
     setLeagueChannel: function (message) {
-        fh.write('LeagueChannel.json', message.channel.id);
+        Channel.set('League', message.channel.id, message.guild.id);
         message.channel.send('This is now the Standard LoL Channel');
     },
 
@@ -104,7 +130,7 @@ module.exports = {
     * @summary sets the TwitchChannel to the current Channel
     */
     setTwitchChannel: function (message) {
-        fh.write('TwitchChannel.json', message.channel.id);
+        Channel.set('Twitch', message.channel.id, message.guild.id);
         message.channel.send('This is now the Standard Twitch Channel');
     },
 
@@ -112,7 +138,7 @@ module.exports = {
     * @summary sets the StandardChannel to the current Channel
     */
     setStandardChannel: function (message) {
-        fh.write('StandardChannel.json', message.channel.id);
+        Channel.set('Standard', message.channel.id, message.guild.id);
         message.channel.send('This is now the Standard Channel for automated Messages');
     },
 
@@ -121,9 +147,9 @@ module.exports = {
      * @returns Settings-Embed to Discord-Chat
      */
     settings: function (message, bot) {
-        
-        var Settings = fh.getSettings(); //Get Settings
-        var Emojis = ['0️⃣','1️⃣','2️⃣','3️⃣','4️⃣','5️⃣','6️⃣','7️⃣','8️⃣','9️⃣','🔟','⬜'];
+
+        var Settings = fh.get('../Files/local/' + message.guild.id + '/settings.json'); //Get Settings
+        var Emojis = ['0️⃣', '1️⃣', '2️⃣', '3️⃣', '4️⃣', '5️⃣', '6️⃣', '7️⃣', '8️⃣', '9️⃣', '🔟', '⬜'];
         var etn = {};
         var msg;
 
@@ -149,26 +175,13 @@ module.exports = {
 
             for (let user of emoji.users) {
 
-                if (Admins.includes(user[0])) {
+                if (Devs.includes(user[0])) { //to be changed to perSever-Check
 
                     msg.clearReactions();
 
                     Settings[etn[emoji._emoji.name]] = !Settings[etn[emoji._emoji.name]];
 
-                    if (Settings[etn[emoji._emoji.name]]) { //Custom Commands for some Settings
-                        switch (etn[emoji._emoji.name]) {
-                            case 'checkForLOLGames':
-                                require('./league').checkForLOLGames(bot);
-                                break;
-                            case 'checkForTwitchStreams':
-                                require('./twitch').checkForStreams(bot);
-                                break;
-                            default:
-                                break;
-                        }
-                    }
-
-                    if (!saveSettings(Settings)) { //Save Settings to settings.json
+                    if (!fh.write('settings.json', Settings, message.guild.id)) { //Save Settings to settings.json
                         message.channel.send('An Error occured while saving Settings');
                     }
 
@@ -192,16 +205,58 @@ module.exports = {
 
             ans.delete(300000); //Delete Message after 5 minutes
         })
+    },
+    /**
+     * @summary Adds a DiscordUser to AdminList for this Server/Guild
+     */
+    addAdmin(message,bot) {
+        let Admins = fh.get('../Files/local/' + message.guild.id + '/Admins.json');
+        let user = message.content.substring(message.content.indexOf(' ') + 9);
+        user = user.substring(4, user.length - 1);
+
+        if (!bot.users.find(u => u.id == user)) {
+            message.channel.send('No User tagged');
+            return;
+        }
+
+        if (Admins.indexOf(user) == -1) {
+            Admins.push(user);
+            fh.write('Admins.json', Admins, message.guild.id);
+            message.channel.send('Added <@!' + user + '> to the AdminList');
+
+        } else {
+            message.channel.send('<@!' + user + '> is already in the AdminList');
+        }
+
+
+    },
+    /**
+     * @summary Removes a DiscordUser to AdminList for this Server/Guild
+     */
+    removeAdmin(message,bot) {
+        let Admins = fh.get('../Files/local/' + message.guild.id + '/Admins.json');
+        let user = message.content.substring(message.content.indexOf(' ') + 12);
+        user = user.substring(4, user.length - 1);
+
+        if (!bot.users.find(u => u.id == user)) {
+            message.channel.send('No User tagged');
+            return;
+        }
+
+        if (Admins.indexOf(user) > -1) {
+            Admins.splice(Admins.indexOf(user),1);
+            fh.write('Admins.json', Admins, message.guild.id);
+            message.channel.send('Removed <@!' + user + '> from the AdminList');
+
+        } else {
+            message.channel.send('<@!' + user + '> is not on the AdminList');
+        }
     }
 }
 
-var Admins = [ //Add DiscordID for AdminAccess
+var Devs = [ //Add DiscordID for DevAccess
     '270929192399536138', //ackhack
     '222398053703876628' //Human Daniel
 ]
 
 var stopvar = false;
-
-function saveSettings(Settings) { //Saves values to settings.json
-    return fh.write('settings.json', Settings);
-}