+++ Commands/FileHandler.js
@@ -13,10 +13,6 @@ module.exports = {
      */
     get: function (Path) {
 
-        if (Path == '../Files/local/settings.json') {
-            return this.getSettings();
-        }
-
         try {
             return require(Path);
         } catch (error) {
@@ -27,62 +23,83 @@ module.exports = {
 
     /**
      * @summary Failsafe fs.writeFileSync(Path,Object)
-     * @param {string} Path Automatically goes to Files/local/FILENAME
+     * @param {string} Path Automatically goes to Files/local/guildID?/FILENAME
+     * @param {Object} Object any Object (maybe use JSON.stringify)
+     * @param {Number} guildID the guildID (message.guild.id)
      */
-    write: function (Path, Object) {
+    write: function (Path, Object, guildID) {
+
         let splitted = Path.split('/');
         let Filename = splitted[splitted.length - 1];
-        return writeFile('Files/local/' + Filename, Object);
-    },
 
-    /**
-     * @summary Improved require('../Files/local/settings.json'), should always be used
-     * @returns {Object} Returns require('../Files/local/settings.json')
-     */
-    getSettings: function () {
-        let set;
-        try {
-            set = require('../Files/local/settings.json'); //Test if settings.json is valid
-
-        } catch (ignored) { //No settings.json
-
-            try {
-                try {   //Try to create the local Folder
-                    fs.mkdirSync('Files/local');
-                } catch (ignored) { }
-
-                fs.writeFileSync('Files/local/settings.json', JSON.stringify(require('../Files/initsettings.json'))); //Create settings.json
-                return require('../Files/local/settings.json');
-
-            } catch (error) {
-                Logger.log(error);
-                return require('../Files/initsettings.json');
+        if (guildID) {
+            if (!fs.existsSync('Files/local/' + guildID)) {
+                fs.mkdirSync('Files/local/' + guildID);
             }
+            return writeFile('Files/local/' + guildID + '/' + Filename, Object);
+
+        } else {
+            return writeFile('Files/local/' + Filename, Object);
         }
-
-        const initset = require('../Files/initsettings.json');
-
-        for (let setting in initset) { //Test for new Settings in initsettings.json
-            if (set[setting] == undefined) {
-                set[setting] = initset[setting];
-            }
-        }
-
-        try { //Write new Settings into settings.json
-            fs.writeFileSync('Files/local/settings.json', JSON.stringify(set));
-        } catch (error) {
-            Logger.log(error);
-        }
-
-        return set;
     },
 
     readdirSync: function (path) { //just calls fs.readDirSync
         return fs.readdirSync(path);
+    },
+
+    /**
+     * @summary inits all settings.json
+     */
+    initSettings(bot) {
+
+        try {   //Try to create the local Folder
+            fs.mkdirSync('Files/local');
+        } catch (ignored) { }
+
+        const initset = require('../Files/initsettings.json');
+
+        for (let guild of bot.guilds) {
+
+            var set;
+            try {
+                set = require('../Files/local/' + guild[0] + '/settings.json'); //Test if settings.json is valid
+
+            } catch (ignored) { //No settings.json
+
+                try {   //Try to create the local Folder
+                    fs.mkdirSync('Files/local/' + guild[0]);
+                } catch (ignored) { }
+
+                try {
+                    fs.writeFileSync('Files/local/' + guild[0] + '/settings.json', JSON.stringify(initset)); //Create settings.json
+                    set = undefined;
+                    continue;
+                } catch (error) {
+                    Logger.log(error);
+                }
+            }
+
+            for (let setting in initset) { //Test for new Settings in initsettings.json
+                if (!set[setting]) {
+                    set[setting] = initset[setting];
+                }
+            }
+
+            try { //Write new Settings into settings.json
+                fs.writeFileSync('Files/local/' + guild[0] + '/settings.json', JSON.stringify(set));
+            } catch (error) {
+                Logger.log(error);
+            }
+            set = undefined;
+        }
     }
 }
 
-function create(Path) {
+function create(Path, guildID) {
+
+    if (!fs.existsSync('Files/local/' + guildID)) {
+        fs.mkdirSync('Files/local/' + guildID);
+    }
 
     let splitted = Path.split('/');
     let Filename = splitted[splitted.length - 1];
@@ -90,16 +107,20 @@ function create(Path) {
 
     switch (Filename) {
 
-        case 'counter.json':
-            File = { "good": 0, "bad": 0, "called": false };
-            break;
+        case 'counter.json': //Only one per Bot
+            writeFile('Files/local/botToken.json', { "good": 0, "bad": 0, "called": false });
+            return;
 
-        case 'botToken.json':
-            File = { "token": "" };
-            break;
+        case 'botToken.json': //Only one per Bot
+            writeFile('Files/local/botToken.json', { "token": "" });
+            return;
 
-        case 'whatToDraw.json':
+        case 'whatToDraw.json': //Only one per Bot
+            writeFile('Files/local/whatToDraw.json', []);
+            return;
+
         case 'Streamers.json':
+        case 'Admins.json':
             File = [];
             break;
 
@@ -110,15 +131,19 @@ function create(Path) {
             break;
 
         case 'names.json':
+        case 'Channels.json':
             File = {};
+            break;
+
+        case 'settings.json': //Should never happen, but just in case
+            File = JSON.stringify(require('../Files/initsettings.json'));
             break;
 
         default:
             File = undefined;
             break;
     }
-    writeFile('Files/local/' + Filename, File);
-    return File;
+    return writeFile('Files/local/' + guildID + '/' + Filename, File) ? File : undefined;
 }
 
 function writeFile(Path, Object) {